requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
gitpython>=3.1
pytest>=8.0
httpx>=0.27